% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAMM.KCCAACCA-package.R
\docType{package}
\name{BEAMM.KCCAACCA-package}
\alias{BEAMM.KCCAACCA}
\alias{BEAMM.KCCAACCA-package}
\title{BEAMM.KCCAACCA-package}
\description{
The BEAMM.KCCAACCA package provides ...
}
\section{BEAMM.KCCAACCA functions}{

The BEAMM.KCCAACCA functions ...
}

\author{
\strong{Maintainer}: Alessandro Beretta \email{alessandro.beretta@usaintlouis.be}

Authors:
\itemize{
  \item Hugues Annoye \email{hugues.annoye@uclouvain.be}
  \item Cedric Heuchenne \email{cedric.heuchenne@uclouvain.be}
  \item Ida-Marie Jensen \email{ida-marie.jensen@usaintlouis.be}
}

}
