% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAMM.statMatch-package.R
\docType{package}
\name{BEAMM.statMatch-package}
\alias{BEAMM.statMatch-package}
\title{BEAMM.statMatch-package}
\description{
The BEAMM.statMatch package provides ...
}
\section{BEAMM.statMatch functions}{

The BEAMM.statMatch functions ...
}

