% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.compute_cramer_von_mises.R
\name{compute_cramer_von_mises}
\alias{compute_cramer_von_mises}
\title{cramer von minsen discrete}
\usage{
compute_cramer_von_mises(
  ACTUAL,
  PRED,
  wt = NULL,
  names.NCV = NULL,
  type = c("mean", "sum", "sumst")
)
}
\arguments{
\item{ACTUAL}{a data frame, actual observations.}

\item{PRED}{a data frame, predictions.}

\item{wt}{a numeric vector, sample weights.}

\item{names.NCV}{a character vector with the names of the non-common variables used for the prediction.}

\item{type}{a string, type of error measure, (either \code{mean}, \code{sum} or \code{sumst}).}
}
\value{
A list,with the values of the objective function.
}
\description{
Compute the value of the compute_pseudo_Rsquare ans make the mean or the sum
}
