% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/utils.compute_cramer_von_mises_for_independance.R
\name{compute_cramer_von_mises_for_independence}
\alias{compute_cramer_von_mises_for_independence}
\title{cramer von minses discrete}
\usage{
compute_cramer_von_mises_for_independence(
  DATA,
  wt = NULL,
  names = NULL,
  type = c("mean", "sum", "sumst"),
  parallel = FALSE,
  nnodes = 10
)
}
\arguments{
\item{DATA}{a data frame with observations.}

\item{wt}{a numeric vector, sample weights for ACTUAL.}

\item{names}{a character vector with the names of the non-common variables used for the prediction.}

\item{type}{a string, type of error measure, (either \code{mean}, \code{sum} or \code{sumst}).}

\item{parallel}{a logical, whether or not use parallel computing.}

\item{nnodes}{an integer, number of parallel session (CPUs).}
}
\value{
A list,with the values of the objective function.
}
\description{
cramer von minses between the multivariate and the indepedence case
}
