% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.compute_multi_Rsquare.R
\name{compute_multi_Rsquare}
\alias{compute_multi_Rsquare}
\title{compute_multi_Rsquare}
\usage{
compute_multi_Rsquare(
  PRED,
  ACTUAL,
  names.NCV,
  weights = NULL,
  type = c("chow", "book", "R", "R_bis")
)
}
\arguments{
\item{PRED}{a data.frame, predictions.}

\item{ACTUAL}{a data.frame, actual observations.}

\item{names.NCV}{a character vector with the names of the variables.}

\item{weights}{a numeric vector, sample weights.}

\item{type}{a character string to indicate which multivariate R square mut be calculated \code{"chow"} for the one of Chow, \code{"book"} for the one of Cohen and \code{R} for the one of Jones.}
}
\value{
A double, the value of the multivariate R square
}
\description{
Compute the multivariate R square
}
