% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.compute_obj_function.R
\name{compute_obj_function}
\alias{compute_obj_function}
\title{compute_obj_function}
\usage{
compute_obj_function(
  ACTUAL,
  PRED,
  wt,
  type = c("wsRMSE", "wRMSE", "wCE"),
  eps = .Machine$double.eps
)
}
\arguments{
\item{ACTUAL}{a numeric matrix, actual observations.}

\item{PRED}{a numeric matrix, predictions.}

\item{wt}{a numeric vector, sample weights.}

\item{type}{a string, type of error measure, (either \code{wsRMSE} or \code{wRMSE}).}

\item{eps}{a double, machine epsilon.}
}
\value{
A double, the value of the objective function.
}
\description{
Compute the value of the objective function (cross-validation), either weighted standardized Root Mean Square Error (\code{wsRMSE}) or weighted Root Mean Square Error (\code{wRMSE}).
}
\details{
The standardized version of the objective function, computed using \code{type = "wsRMSE"}, enables us to deal with the fact the continuous variables are measured on different scales.
}
