% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.compute_pseudo_Rsquare_new.R
\name{compute_pseudo_Rsquare_new}
\alias{compute_pseudo_Rsquare_new}
\title{ccompute_pseudo_Rsquare}
\usage{
compute_pseudo_Rsquare_new(ACTUAL, PRED, wt, type = c("mean", "sum"))
}
\arguments{
\item{ACTUAL}{a numeric matrix, actual observations.}

\item{PRED}{a numeric matrix, predictions.}

\item{wt}{a numeric vector, sample weights.}

\item{type}{a string, type of error measure, (either \code{mean} or \code{sum}).}
}
\value{
A double, the value of the objective function.
}
\description{
Compute the value of the compute_pseudo_Rsquare ans make the mean or the sum
}
\details{
The standardized version of the objective function, computed using \code{type = "wsRMSE"}, enables us to deal with the fact the continuous variables are measured on different scales.
}
