% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.df2mtx.R
\name{df2mtx}
\alias{df2mtx}
\title{df2mtx}
\usage{
df2mtx(df, n_1_levels = FALSE)
}
\arguments{
\item{df}{a data.frame.}

\item{n_1_levels}{a logical value, if TRUE keep only \eqn{n-1} dummies for each factor.}
}
\value{
A numeric matrix.
}
\description{
Transform a data.frame into a matrix (factors are transformed into \eqn{n} dummy variables, where \eqn{n} is the number of levels)
}
