% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.ebvcdfw3.R
\name{ebvcdfw3}
\alias{ebvcdfw3}
\title{cramer von minsen discrete}
\usage{
ebvcdfw3(x, y, z, wt, ..., data)
}
\arguments{
\item{x}{Variable x}

\item{y}{Variable y}

\item{z}{Variable z}

\item{wt}{a numeric vector, sample weights.}

\item{...}{ellipsis}

\item{data}{data}
}
\value{
a function
}
\description{
Compute the value of the compute_pseudo_Rsquare ans make the mean or the sum
}
