% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.ebvcdfwall.R
\name{ebvcdfwall}
\alias{ebvcdfwall}
\title{cramer von minsen discrete}
\usage{
ebvcdfwall(x, wt, ..., data)
}
\arguments{
\item{x}{Dataframe}

\item{wt}{a numeric vector, sample weights.}

\item{...}{ellipsis}

\item{data}{data}
}
\value{
a function
}
\description{
Compute the value of the compute_pseudo_Rsquare ans make the mean or the sum
}
