% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactMatch.R
\name{exactMatch}
\alias{exactMatch}
\title{Exact Matching: SOM Tuning}
\usage{
exactMatch(df.rec, df.don, names.CV, names.NCV)
}
\arguments{
\item{df.rec}{a data.frame, the recipient data set.}

\item{df.don}{a data.frame, the donor data set.}

\item{names.CV}{a character vector with the names of the common variables used for statistical matching.}

\item{names.NCV}{a character vector with the names of the non-common variables used for statistical matching.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
What is doing this function?
}
\details{
DETAILS
}
