% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.min_dist.R
\name{min_dist}
\alias{min_dist}
\title{Minimal distance between two matrice.}
\usage{
min_dist(mtx1, mtx2)
}
\arguments{
\item{mtx1}{a numeric matrix.}

\item{mtx2}{a numeric matrix.}
}
\value{
A numeric vect
}
\description{
min_dist
}
