% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.mtx2df.R
\name{mtx2df}
\alias{mtx2df}
\title{mtx2df}
\usage{
mtx2df(mtx, var_names)
}
\arguments{
\item{mtx}{a numeric matrix.}

\item{var_names}{a character vector, for the identification of the dummy variables belonging to a given factor.}
}
\value{
A data.frame.
}
\description{
Transform a matrix into a data.frame (dummy variables variables are transformed into factors)
}
