% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.normalize_mtx.R
\name{normalize_mtx}
\alias{normalize_mtx}
\title{normalize_mtx}
\usage{
normalize_mtx(mtx, min = NULL, max = NULL)
}
\arguments{
\item{mtx}{a numeric matrix.}

\item{min}{a numeric vector.}

\item{max}{a numeric vector.}
}
\value{
A numeric matrix.
}
\description{
Scale a matrix (min-max) such that column minimums and maximums are equal to 0 and 1, respectively.
}
