% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.normalize_mtx_inv.R
\name{normalize_mtx_inv}
\alias{normalize_mtx_inv}
\title{normalize_mtx_inv}
\usage{
normalize_mtx_inv(mtx_norm, min, max)
}
\arguments{
\item{mtx_norm}{a matrix object.}

\item{min}{a numeric vector.}

\item{max}{a numeric vector.}
}
\value{
A matrix object.
}
\description{
Transform a matrix from scaled (min-max) to unscaled.
}
\details{
DETAILS
}
