% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.scale_mtx.R
\name{scale_mtx}
\alias{scale_mtx}
\title{scale_mtx}
\usage{
scale_mtx(mtx, wt = NULL, center = NULL, scale = NULL)
}
\arguments{
\item{mtx}{a numeric matrix.}

\item{wt}{a numeric vector, sample weights.}

\item{center}{a numeric vector.}

\item{scale}{a numeric vector.}
}
\value{
A numeric matrix.
}
\description{
Scale a matrix (z-score) such that column means and standard deviations are equal to 0 and 1, respectively.
}
\details{
If \code{center} or/and \code{scale} are equal to \code{NULL}, the function compute the column means and standard deviations.
}
