% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.scale_mtx_inv.R
\name{scale_mtx_inv}
\alias{scale_mtx_inv}
\title{scale_mtx_inv}
\usage{
scale_mtx_inv(mtx_scaled, center, scale)
}
\arguments{
\item{mtx_scaled}{a numeric matrix.}

\item{center}{a numeric vector.}

\item{scale}{a numeric vector.}
}
\value{
A numeric matrix.
}
\description{
Transform a matrix from scaled (z-score) to unscaled.
}
