% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.simplePDP.R
\name{simple_PDP}
\alias{simple_PDP}
\title{simple_PDP}
\usage{
simple_PDP(
  df_don,
  names.CV,
  names.NCV,
  names.ZC,
  names_weights,
  grid_point,
  method,
  options,
  input,
  prediction
)
}
\arguments{
\item{df_don}{a data.frame}

\item{names.CV}{a character vector with the names of the common variables used for statistical matching.}

\item{names.NCV}{a character vector with the names of the non-common variables used for statistical matching.}

\item{names.ZC}{character vector with the names of the zero constraint variables used for statistical matching.}

\item{names_weights}{a numeric vector, the sample weights of the donor data set.}

\item{grid_point}{number of grid points for estimation}

\item{method}{a character with the name of statistical matching method.}

\item{options}{a list, supplied using the function \code{statMatch.MLP.options()}, containing all the options.}

\item{input}{a character with the name of one or two common variables}

\item{prediction}{a character with the name of one non-common variable}
}
\value{
A plot, ...
}
\description{
Plot partial dependence plot for one or two variables without kernel
}
