% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMatch.bootstrap.R
\name{statMatch.bootstrap}
\alias{statMatch.bootstrap}
\title{Statistical Matching using bootstrap}
\usage{
statMatch.bootstrap(
  df.rec,
  df.don,
  don.weights = NULL,
  boot_fold = 10,
  length_boot = 1000,
  opts = NULL,
  n_cores = 3,
  print.boot = TRUE,
  names.CV = NULL,
  names.NCV = NULL,
  zero.constraints = NULL,
  method = method
)
}
\arguments{
\item{df.rec}{a data.frame, the recipient data set.}

\item{df.don}{a data.frame, the donor data set.}

\item{don.weights}{a numeric vector, the sample weights of the donor data set.}

\item{boot_fold}{a positive integer, with number of folds}

\item{length_boot}{a positive integer, with length of each bootstrap fold}

\item{opts}{a list, supplied using a function \code{statMatch.***.options()}, containing all the options.}

\item{n_cores}{a positive integer, with number of cores}

\item{print.boot}{a logical value, print tuned parameter or not}

\item{names.CV}{a character vector with the names of the common variables used for statistical matching.}

\item{names.NCV}{a character vector with the names of the non-common variables used for statistical matching.}

\item{zero.constraints}{a character vector with the names of the non-common variables subject to the zero constraint.}

\item{method}{a character vector with the name of statmatch method that have to be used}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Make a boostrap statistical matching using the function statMatch
}
\details{
DETAILS
}
