% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMatch.KCCA.one.phase.options.R
\name{statMatch.KCCA.one.phase.options}
\alias{statMatch.KCCA.one.phase.options}
\title{Statistical Matching: KCCA options}
\usage{
statMatch.KCCA.one.phase.options(
  print.details = TRUE,
  print.details_tuning = FALSE,
  scaling = c("z-score", "min-max", "no"),
  tuning_type = c("two h", "one h", "random"),
  tuning_type_bandwith = c("one h", "two h"),
  type_predict = c("loop", "matrix"),
  exactMatch = FALSE,
  names.CV_Cat_prio = NULL,
  n_fold = 5L,
  par = FALSE,
  par_fold = FALSE,
  nc = 2L,
  d = 2L,
  rot = FALSE,
  p0_kernel_predict = c("gauss", "alea", "epan"),
  p0_kernel_predict_cat = c("alea", "epan", "gauss"),
  p0_objmethod = c("wsRMSE", "wRMSE", "wMCR", "wCE"),
  p0_h = NULL,
  p0_hmax = 1,
  p0_hmin = 0.01,
  p0_hstep = 0.1,
  p0_h_cat = NULL,
  p0_hmax_cat = NULL,
  p0_hmin_cat = NULL,
  p0_hstep_cat = NULL,
  p0_hx = NULL,
  p0_hxmax = 1,
  p0_hxmin = 0.01,
  p0_hxstep = 0.25,
  p0_hy = NULL,
  p0_hymax = NULL,
  p0_hymin = NULL,
  p0_hystep = NULL,
  p0_g = NULL,
  p0_gmax = 0.5,
  p0_gmin = 0.01,
  p0_gstep = 0.1,
  p0_n_combs = 10L,
  p0_man_params = NULL
)
}
\arguments{
\item{print.details}{Logical Value, print details or not (in the tuning part only print the value of the objective function for each parameters combination).}

\item{print.details_tuning}{Logical Value, if true print all the details in the tuning part.}

\item{scaling}{a character string, the name of the method used to scale the variables in the receiver and donor data sets: \code{"z-score"}, \code{"min-max"} or \code{"no"} (no scaling).}

\item{tuning_type}{a character string, the method used to find the optimal tuning parameters: \code{"random"}, \code{"two h"} or \code{"one h"}}

\item{type_predict}{a character string, Type of prediction \code{"matrix"} or \code{"loop"}.}

\item{exactMatch}{a logical, indicate if an exact matching has to be performed or not.}

\item{names.CV_Cat_prio}{a character vector, Name of common categorical variable that have to match.}

\item{n_fold}{a positive integer, number of folds used for the cross-validation of the tuning parameters.}

\item{par}{a logical, if TRUE doing parallelisation during the tuning phase}

\item{nc}{a positive integer, number of core to use during the tuning phase}

\item{d}{a positive integer, number of latent variable used in CCA.}

\item{rot}{a logical, if TRUE the bandwidth is h multiply by the variance.}

\item{p0_kernel_predict}{a character string, Type of kernel use for prediction for continuous variable.}

\item{p0_kernel_predict_cat}{a character string, Type of kernel use for prediction for categorical variable.}

\item{p0_objmethod}{a character string, Method using for the tuning.}

\item{p0_h}{a numeric vector, vector of value for h, the bandwidth of the Matching kernel for categorical variable.}

\item{p0_hmax}{a numeric value, maximal value for h, the bandwidth of the Matching kernel for categorical variable.}

\item{p0_hmin}{a numeric value, minimal value for h, the bandwidth of the Matching kernel for categorical variable.}

\item{p0_hstep}{a numeric value, step for the grid of h for categorical variable.}

\item{p0_hx}{a numeric vector, vector of value for h, the bandwidth of the KCCA kernel for categorical variable.}

\item{p0_hxmax}{a numeric value, maximal value for h, the bandwidth of the KCCA kernel for categorical variable.}

\item{p0_hxmin}{a numeric value, minimal value for h, the bandwidth of the KCCA kernel for categorical variable.}

\item{p0_hxstep}{a numeric value,  step for the grid of h for categorical variable.}

\item{p0_hy}{a numeric vector, vector of value for h, the bandwidth of the KCCA kernel for categorical variable.}

\item{p0_hymax}{a numeric value, maximal value for h, the bandwidth of the KCCA kernel for categorical variable.}

\item{p0_hymin}{a numeric value, minimal value for h, the bandwidth of the KCCA kernel for categorical variable.}

\item{p0_hystep}{a numeric value, step for the grid of h for categorical variable.}

\item{p0_g}{a numeric vector, vector of value for g, the regularization parameter for categorical variable.}

\item{p0_gmax}{a numeric value, maximal value for g, the regularization parameter for categorical variable.}

\item{p0_gmin}{a numeric value, minimal value for g, the regularization parameter for categorical variable.}

\item{p0_gstep}{a numeric value, mtep for the grid of g for categorical variable.}

\item{p0_n_combs}{a positive integer, number of random combinations of tuning parameters generated in phase 1 (used only when \code{tuning_type = "random"}).}

\item{p0_man_params}{a data.frame, a manual way to specify the grid of tuning parameters in phase 1.}
}
\value{
A list with all the options required by the statistical matching algorithm based on KCCA.
}
\description{
Function to create the options for KCCA.
}
\details{
DETAILS
}
