% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMatch.PCA.options.R
\name{statMatch.PCA.options}
\alias{statMatch.PCA.options}
\title{Statistical Matching: PCA options}
\usage{
statMatch.PCA.options(
  print.details = TRUE,
  print.details_tuning = FALSE,
  scaling = c("z-score", "min-max", "no"),
  tuning_type = c("grid", "random"),
  type_predict = c("loop", "matrix"),
  exactMatch = FALSE,
  names.CV_Cat_prio = NULL,
  n_fold = 5L,
  d = 2L,
  rot = FALSE,
  p1_kernel_predict = c("alea", "epan", "gauss"),
  p1_kernel_predict_tuning = NULL,
  p2_kernel_predict = c("gauss", "epan", "alea"),
  p1_lat_X_min = 5L,
  p1_lat_X_max = 10L,
  p1_lat_X_step = 1L,
  p2_lat_X_min = 5L,
  p2_lat_X_max = 10L,
  p2_lat_X_step = 1L,
  p1_lat_Y_min = 5L,
  p1_lat_Y_max = 10L,
  p1_lat_Y_step = 1L,
  p2_lat_Y_min = 5L,
  p2_lat_Y_max = 10L,
  p2_lat_Y_step = 1L,
  p1_objmethod = c("wRMSE", "wsRMSE", "wMCR", "wCE"),
  p2_objmethod = c("wsRMSE", "wRMSE", "wCE"),
  p1_h = NULL,
  p1_hmax = 1,
  p1_hmin = 0.01,
  p1_hstep = 0.1,
  p2_h = NULL,
  p2_hmax = 1,
  p2_hmin = 0.01,
  p2_hstep = 0.1,
  p1_n_combs = 10L,
  p2_n_combs = 10L,
  p1_man_params = NULL,
  p2_man_params = NULL
)
}
\arguments{
\item{print.details}{a logical, print details or not (in the tuning part only print the value of the objective function for each parameters combination)}

\item{print.details_tuning}{a logical, if true print all the details in the tuning part}

\item{scaling}{a character string, the name of the method used to scale the variables in the receiver and donor data sets: \code{"z-score"}, \code{"min-max"} or \code{"no"} (no scaling).}

\item{tuning_type}{a character string, the method used to find the optimal tuning parameters: \code{"random"}, \code{"two h"} or \code{"one h"}}

\item{type_predict}{a character string, Type of prediction \code{"matrix"} or \code{"loop"}.}

\item{exactMatch}{a logical, indicate if an exact matching has to be performed or not.}

\item{names.CV_Cat_prio}{a character vector, Name of common categorical variable that have to match.}

\item{n_fold}{a positive integer, number of folds used for the cross-validation of the tuning parameters.}

\item{d}{a positive integer, number of latent variable used in CCA.}

\item{rot}{a logical, if TRUE the bandwidth is h multiply by the variance.}

\item{p1_kernel_predict}{a character string, Type of kernel use for prediction for categorical variable.}

\item{p1_kernel_predict_tuning}{a character string, Type of kernel use for the tuning for categorical variable by default the same than kernel_predict.}

\item{p2_kernel_predict}{a character string, Type of kernel use for prediction for continuous variable.}

\item{p1_lat_X_min}{a positive integer, minimum size of the latent space of the common variables in phase 1.}

\item{p1_lat_X_max}{a positive integer, maximum size of the latent space of the common variables in phase 1.}

\item{p1_lat_X_step}{a positive integer, step size of the latent space of the common variables in phase 1.}

\item{p2_lat_X_min}{a positive integer, minimum size of the latent space of the common variables in phase 2.}

\item{p2_lat_X_max}{a positive integer, maximum size of the latent space of the common variables in phase 2.}

\item{p2_lat_X_step}{a positive integer, step size of the latent space of the common variables in phase 2.}

\item{p1_lat_Y_min}{a positive integer, minimum size of the latent space of the non-common variables in phase 1.}

\item{p1_lat_Y_max}{a positive integer, maximum size of the latent space of the non-common variables in phase 1.}

\item{p1_lat_Y_step}{a positive integer, step size of the latent space of the non-common variables in phase 1.}

\item{p2_lat_Y_min}{a positive integer, minimum size of the latent space of the non-common variables in phase 2.}

\item{p2_lat_Y_max}{a positive integer, maximum size of the latent space of the non-common variables in phase 2.}

\item{p2_lat_Y_step}{a positive integer, step size of the latent space of the non-common variables in phase 2.}

\item{p1_objmethod}{a character string, the method using for the tuning.}

\item{p2_objmethod}{a character string, the method using for the tuning.}

\item{p1_h}{a numeric vector, the vector of value for h, the bandwidth of the Matching kernel for categorical variable.}

\item{p1_hmax}{a numeric value, the maximal value for h, the bandwidth of the Matching kernel for categorical variable.}

\item{p1_hmin}{a numeric value, the minimal value for h, the bandwidth of the Matching kernel for categorical variable.}

\item{p1_hstep}{a numeric value, a step for the grid of h for categorical variable.}

\item{p2_h}{a numeric vector, the vector of value for h, the bandwidth of the Matching kernel for continuous variable.}

\item{p2_hmax}{a numeric value, the maximal value for h, the bandwidth of the Matching kernel for continuous variable.}

\item{p2_hmin}{a numeric value, the minimal value for h, the bandwidth of the Matching kernel for continuous variable.}

\item{p2_hstep}{a numeric value, a step for the grid of h for continuous variable.}

\item{p1_n_combs}{a positive integer, number of random combinations of tuning parameters generated in phase 1 (used only when \code{tuning_type = "random"}).}

\item{p2_n_combs}{a positive integer, number of random combinations of tuning parameters generated in phase 2 (used only when \code{tuning_type = "random"}).}

\item{p1_man_params}{a data.frame, a manual way to specify the grid of tuning parameters in phase 1.}

\item{p2_man_params}{a data.frame, a manual way to specify the grid of tuning parameters in phase 2.}
}
\value{
A list with all the options required by the statistical matching algorithm based on PCA-CCA.
}
\description{
Function to create the options for PCA
}
\details{
DETAILS
}
