% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMatch.R
\name{statMatch}
\alias{statMatch}
\title{statMatch}
\usage{
statMatch(
  df.rec,
  df.don,
  names.CV = NULL,
  names.NCV = NULL,
  zero.constraints = NULL,
  don.weights = NULL,
  opts = NULL,
  method = c("KCCA", "superOM", "ACCA", "CCA", "PCA", "MLP")
)
}
\arguments{
\item{df.rec}{a data.frame, the recipient data set.}

\item{df.don}{a data.frame, the donor data set.}

\item{names.CV}{a character vector with the names of the common variables used for statistical matching.}

\item{names.NCV}{a character vector with the names of the non-common variables used for statistical matching.}

\item{zero.constraints}{a character vector with the names of the non-common variables subject to the zero constraint.}

\item{don.weights}{a numeric vector, the sample weights of the donor data set.}

\item{opts}{a list containing all the options supplied by the function \code{statMatch.KCCA.options()}, \code{statMatch.superOM.options()}, \code{statMatch.ACCA.options()}, ..., depending on the selected method.}

\item{method}{a string, the method used to perform statistical matching (either \code{KCCA}, \code{superOM}, \code{ACCA}, \code{CCA}, \code{PCA} or \code{MMLP}).}
}
\value{
The output is a list containing the following elements:
  \describe{
    \item{match}{Final matched data set (where the predicted zero constraint are applied to the non-common continuous variable(s), if any.)}
    \item{final}{Final matched data set (where the predicted zero constraint are applied to the non-common continuous variable(s), if any.) and where we keep the variables that are not used as common variables in the receiver}
    \item{phase1}{The results of phase1 of the algorthm. A list composed by:}
    \itemize{
         \item{\code{tune}, a data.frame with the combination of hyper-parameters used for cross-validation and the values of the corresponding objective function.}
         \item{\code{fit}, result of the fitted KCCA, CCA, ACCA, super-OM, ... with the best combination of hyper-parameters.}
         \item{\code{df.match.don}}
         \item{\code{df.match}}
         \item{\code{sum_dist_comp_donor}}
    }
    \item{phase2}{The results of phase2 of the algorithm. A list containing:}
    \itemize{
         \item{\code{tune}, a data.frame with the combination of hyper-parameters used for cross-validation and the values of the corresponding objective function.}
         \item{\code{fit}, result of the fitted KCCA, CCA, ACCA, super-OM, ... with the best combination of hyper-parameters.}
         \item{\code{df.match.don}}
         \item{\code{df.match}}
    }
    \item{opts}{options provided to the statistical matching algorithm.}
  }
}
\description{
Main function to perform statistical matching
}
\details{
DETAILS
}
