% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMatch.superSOM.options.R
\name{statMatch.superOM.options}
\alias{statMatch.superOM.options}
\title{statMatch.superOM.options}
\usage{
statMatch.superOM.options(
  print.details = TRUE,
  scaling = c("z-score", "min-max", "no"),
  tuning_type = c("grid", "random"),
  p1_rand_combs = 10L,
  p2_rand_combs = 10L,
  p1_square_grid = TRUE,
  p2_square_grid = TRUE,
  p1_dim_min = 2L,
  p2_dim_min = 2L,
  p1_dim_max = 20L,
  p2_dim_max = 20L,
  p1_dim_step = 1L,
  p2_dim_step = 1L,
  p1_wCV_min = 0,
  p2_wCV_min = 0,
  p1_wCV_max = 1,
  p2_wCV_max = 1,
  p1_wCV_step = 0.1,
  p2_wCV_step = 0.1,
  n_fold = 5L,
  n_nodes = 1L,
  topo = c("hexagonal", "rectangular"),
  neighbourhood.fct = c("gaussian", "bubble"),
  toroidal = FALSE,
  rlen = 250L,
  dist.fcts = c("euclidean"),
  maxNA.fraction = 0.9,
  seed_phase1 = NULL,
  seed_phase2 = NULL,
  man_params_1 = NULL,
  man_params_2 = NULL
)
}
\arguments{
\item{print.details}{a logical, whether or not details about the different procedures involved in the statistical matching are printed on the terminal.}

\item{scaling}{a character string, the name of the method used to scale the variables in the receiver and donor data sets: \code{"z-score"}, \code{"min-max"} or \code{"no"} (no scaling).}

\item{tuning_type}{a character string, the method used to find the optimal tuning parameters: either \code{"grid"} or \code{"random"} search.}

\item{p1_rand_combs}{a positive integer, number of random combinations of tuning parameters generated in phase 1 (used only when \code{tuning_type = "random"}).}

\item{p2_rand_combs}{a positive integer, number of random combinations of tuning parameters generated in phase 2 (used only when \code{tuning_type = "random"}).}

\item{p1_square_grid}{a logical, whether or not a superOM with a square grid is used in phase 1.}

\item{p2_square_grid}{a logical, whether or not a superOM with a square grid is used in phase 2.}

\item{p1_dim_min}{a positive integer, minimal length of the x and y axis of the superOM grid in phase 1.}

\item{p2_dim_min}{a positive integer, minimal length of the x and y axis of the superOM grid in phase 2.}

\item{p1_dim_max}{a positive integer, maximal length of the x and y axis of the superOM grid in phase 1.}

\item{p2_dim_max}{a positive integer, maximal length of the x and y axis of the superOM grid in phase 2.}

\item{p1_dim_step}{a positive integer, step size of the x and y axis of the superOM grid in phase 1 (used only when \code{tuning_type = "grid"}).}

\item{p2_dim_step}{a positive integer, step size of the x and y axis of the superOM grid in phase 2 (used only when \code{tuning_type = "grid"}).}

\item{p1_wCV_min}{a double, minimal value of the weight of the superOM layer containing the common variables in phase 1.}

\item{p2_wCV_min}{a double, minimal value of the weight of the superOM layer containing the common variables in phase 2.}

\item{p1_wCV_max}{a double, maximal value of the weight of the superOM layer containing the common variables in phase 1.}

\item{p2_wCV_max}{a double, maximal value of the weight of the superOM layer containing the common variables in phase 2.}

\item{p1_wCV_step}{a double, step size of the weight of the superOM layer containing the common variables in phase 1 (used only when \code{tuning_type = "grid"}).}

\item{p2_wCV_step}{a double, step size of the weight of the superOM layer containing the common variables in phase 2 (used only when \code{tuning_type = "grid"}).}

\item{n_fold}{a positive integer, number of folds used for the cross-validation of the tuning parameters.}

\item{n_nodes}{a positive integer, number of parallel sessions (CPUs) used for the cross-validation of the tuning parameters.}

\item{topo}{a character string, topology of the superOM grid, either \code{"hexagonal"} or \code{"rectangular"}.}

\item{neighbourhood.fct}{a character string, neighbourhood function used to train the superOM, either \code{"gaussian"} or \code{"bubble"}.}

\item{toroidal}{a logical, whether the superOM grid is toroidal or not.}

\item{rlen}{a positive integer, the number of times the complete data set will be presented to the network when the training of the superOM.}

\item{dist.fcts}{a character string, distance function used for the individual data layers (only \code{"euclidean"}).}

\item{maxNA.fraction}{a double, the maximal fraction of values that may be NA to prevent the row to be removed when training the superOM.}

\item{seed_phase1}{a positive integer, the seed for random number generation in phase 1 (a random value if \code{NULL}).}

\item{seed_phase2}{a positive integer, the seed for random number generation in phase 2 (a random value if \code{NULL}).}

\item{man_params_1}{a data.frame, a manual way to specify the grid of tuning parameters in phase 1.}

\item{man_params_2}{a data.frame, a manual way to specify the grid of tuning parameters in phase 2.}
}
\value{
A list with the options required to the statistical matching algorithm using superOM.
}
\description{
Create a list with the options for the statistical matching algorithm based on superOM
}
