% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statMatch.superSOM.stabilizer.R
\name{statMatch.superOM.stabilizer}
\alias{statMatch.superOM.stabilizer}
\title{statMatch.superOM.stabilizer}
\usage{
statMatch.superOM.stabilizer(
  df.rec,
  df.don,
  names.CV = NULL,
  names.NCV = NULL,
  zero.constraints = NULL,
  don.weights = NULL,
  exactMatch = FALSE,
  opts = NULL,
  n_trials = 5
)
}
\arguments{
\item{df.rec}{a data.frame, the recipient data set.}

\item{df.don}{a data.frame, the donor data set.}

\item{names.CV}{a character vector with the names of the common variables used for statistical matching.}

\item{names.NCV}{a character vector with the names of the non-common variables used for statistical matching.}

\item{zero.constraints}{a character vector with the names of the non-common variables subject to the zero constraint.}

\item{don.weights}{a numeric vector, the sample weights of the donor data set.}

\item{exactMatch}{a logical, whether or not performing exact matching (before statistical matching).}

\item{opts}{a list, supplied using the function \code{statMatch.superOM.options()}, containing all the options.}

\item{n_trials}{an integer, the number of trials used to stabilize the results.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Main function to perform statistical matching using superOM
}
\details{
DETAILS
}
